api/requirements.txt
anyio==3.6.1
appdirs==1.4.4
asgiref==3.5.2
attrs==22.2.0
beautifulsoup4==4.11.1
Brotli==1.0.9
bs4==0.0.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cssselect==1.1.0
dnspython==2.2.1
email-validator==1.2.1
exceptiongroup==1.1.0
fake-useragent==0.1.11
fastapi==0.78.0
h11==0.13.0
httptools==0.4.0
humps==0.2.2
idna==3.3
importlib-metadata==4.12.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
MechanicalSoup==1.1.0
mutagen==1.46.0
orjson==3.7.1
packaging==23.0
parse==1.19.0
Pillow==9.1.1
pluggy==1.0.0
pycryptodomex==3.17
pydantic==1.9.1
pyee==8.2.2
pyhumps==3.8.0
pyppeteer==1.0.2
pyquery==1.4.3
pytest==7.2.2
python-decouple==3.6
python-dotenv==0.20.0
python-multipart==0.0.5
PyYAML==6.0
qrcode==7.3.1
requests==2.27.1
requests-html==0.10.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
starlette==0.19.1
tomli==2.0.1
tqdm==4.64.0
twitter-scraper==0.4.4
typing_extensions==4.2.0
ujson==5.3.0
urllib3==1.26.9
uvicorn==0.17.6
w3lib==1.22.0
watchgod==0.8.2
websockets==10.3
youtube-dl==2021.12.17
yt-dlp==2022.11.11
zipp==3.8.1
